[Feat]채팅방 레이아웃 분리(AI/mento)
