+++ src/app/(stack)/chat/[id]/components/Chat.tsx
