+++ src/app/(stack)/chat/[id]/components/HelpPage.tsx
